Reorganize extensions API and add pagination
