+++ recipe-server/normandy/studies/api/v2/serializers.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+
+from normandy.studies.models import Extension
+
+
+class ExtensionSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Extension
+        fields = [
+            'id',
+            'name',
+            'xpi',
+        ]
