+++ recipe-server/normandy/studies/api/v2/views.py
@@ -1,18 +1,9 @@
-from rest_framework import permissions, serializers, viewsets
+from rest_framework import permissions, viewsets
 
 from normandy.base.api.mixins import CachingViewsetMixin
 from normandy.base.api.permissions import AdminEnabledOrReadOnly
+from normandy.studies.api.v2.serializers import ExtensionSerializer
 from normandy.studies.models import Extension
-
-
-class ExtensionSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Extension
-        fields = [
-            'id',
-            'name',
-            'xpi',
-        ]
 
 
 class ExtensionViewSet(CachingViewsetMixin, viewsets.ModelViewSet):
@@ -22,4 +13,3 @@ class ExtensionViewSet(CachingViewsetMixin, viewsets.ModelViewSet):
         AdminEnabledOrReadOnly,
         permissions.DjangoModelPermissionsOrAnonReadOnly,
     ]
-    pagination_class = None
